+++ dengine/game/baseLevel/BaseLevel.h
@@ -1,14 +1,14 @@
 #ifndef DENGINE_BASELEVEL_H
 #define DENGINE_BASELEVEL_H
 
-#include <SDL.h>
-#include "../include/Utils/Serializer.h"
-#include "../include/System/GameLevel.h"
-#include "../include/System/GameStates.h"
-#include "../include/System/Game.h"
-#include "Player.h"
-#include "../include/System/Sound.h"
-#include "../include/System/AudioManager.h"
+#include "SDL.h"
+#include "../../include/Utils/Serializer.h"
+#include "../../include/System/GameLevel.h"
+#include "../../include/System/GameStates.h"
+#include "../../include/System/Game.h"
+#include "../Player.h"
+#include "../../include/System/Sound.h"
+#include "../../include/System/AudioManager.h"
 
 
 using namespace dengine;
