+++ dengine/game/enemies/LeafMan.h
@@ -3,10 +3,10 @@
 #ifndef DENGINE_LEAFMAN_H
 #define DENGINE_LEAFMAN_H
 
-#include "../include/System/Component.h"
-#include "../include/System/GameObject.h"
-#include "../include/Components/Sprite.h"
-#include "./Player.h"
+#include "../../include/System/Component.h"
+#include "../../include/System/GameObject.h"
+#include "../../include/Components/Sprite.h"
+#include "../Player.h"
 
 using namespace dengine;
