+++ dengine/game/hud.cpp
@@ -1,0 +1,7 @@
+//
+// Created by ernes on 7/6/2023.
+//
+
+#include "hud.h"
+
+using namespace DemoGame;
